### 2022-05-11 #4384 @LucyMcGowan
Thank you! For reviewer suggestions, someone with R and epidemiology experience would be ideal. A few suggestions: masalmon, tgerke, noamross, corinne-riddell, jenniferthompson, darrendahley
