### 2022-05-18 #4384 @fabian-s
:wave: @wxwx1993 @tedinburgh @stmcg

would you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html
