### 2022-05-28 #4384 @fabian-s
:wave: @csinva @AvianaGlobal @wxwx1993 @tedinburgh @corinne-riddell @tgerke

would any of you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html
