### 2022-06-03 #4384 @fabian-s
:wave: @csinva @AvianaGlobal @gcskoenig @Alex-Markham

would any of you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html
