### 2022-06-03 #4384 @Alex-Markham
@fabian-s I'd be willing if it were in Python, but I think I'm unfortunately not familiar enough with R
