### 2022-06-08 #4384 @fabian-s
:wave: @gcskoenig @dandls  @csinva @AvianaGlobal

would any of you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html
