### 2022-06-08 #4384 @fabian-s
> @LucyMcGowan sorry for the delay, as you can see I'm having a hard time finding reviewers...
>
> To speed things up once we actually get going, in the meantime:
>
>     1. please add the missing DOIs that the bot is complaining about
>
>     2. consider shortening your paper -- it's currently more than double the typical/recommended length for JOSS papers (~ 500 - 1000 words, yours is 2500). For example, you could consider moving the "tipping point" example into a package vignette and link the vignette from your paper. I'm also not convinced the full listing of all package functions is necessary in the article.

reposted so it does not get lost here.
