### 2022-06-08 e5d5101
update deprecated function
