### 2022-06-08 ed0958a
update deprecated function
