+++ paper.bib
@@ -1,587 +1,33 @@
-@article{murff2018metformin,
-  title={Metformin use and incidence cancer risk: evidence for a selective protective effect against liver cancer},
-  author={Murff, Harvey J and Roumie, Christianne L and Greevy, Robert A and Hackstadt, Amber J and McGowan, Lucy ED’Agostino and Hung, Adriana M and Grijalva, Carlos G and Griffin, Marie R},
-  journal={Cancer Causes \& Control},
-  volume={29},
-  number={9},
-  pages={823--832},
-  year={2018},
-  publisher={Springer}
-}
-
-@article{rubin1974estimating,
-  title={Estimating causal effects of treatments in randomized and nonrandomized studies.},
-  author={Rubin, Donald B},
-  journal={Journal of educational Psychology},
-  volume={66},
-  number={5},
-  pages={688},
-  year={1974},
-  publisher={American Psychological Association}
-}
-@article{cinelli2020making,
-  title={Making sense of sensitivity: Extending omitted variable bias},
-  author={Cinelli, Carlos and Hazlett, Chad},
-  journal={Journal of the Royal Statistical Society: Series B (Statistical Methodology)},
-  volume={82},
-  number={1},
-  pages={39--67},
-  year={2020},
-  publisher={Wiley Online Library},
-  addendum={This paper describes how to parameterize sensitivity analyses using $R^2$ and introduces the robustness value.}
-}
-
-@article{pun2019caring,
-  title={Caring for critically ill patients with the ABCDEF bundle: results of the ICU liberation collaborative in over 15,000 adults},
-  author={Pun, Brenda T and Balas, Michele C and Barnes-Daly, Mary Ann and Thompson, Jennifer L and Aldrich, J Matthew and Barr, Juliana and Byrum, Diane and Carson, Shannon S and Devlin, John W and Engel, Heidi J and others},
-  journal={Critical care medicine},
-  volume={47},
-  number={1},
-  pages={3},
-  year={2019},
-  publisher={NIH Public Access}
-}
-@article{murff2018metformin,
-  title={Metformin use and incidence cancer risk: evidence for a selective protective effect against liver cancer},
-  author={Murff, Harvey J and Roumie, Christianne L and Greevy, Robert A and Hackstadt, Amber J and McGowan, Lucy ED’Agostino and Hung, Adriana M and Grijalva, Carlos G and Griffin, Marie R},
-  journal={Cancer Causes \& Control},
-  volume={29},
-  number={9},
-  pages={823--832},
-  year={2018},
-  publisher={Springer}
-}
-@Manual{tipr,
-    title = {tipr: Tipping Point Analyses},
-    author = {Lucy D'Agostino McGowan},
-    year = {2022},
-    note = {R package version 0.4.0},
-  }
-@article{vanderweele2020optimal,
-  title={Optimal approximate conversions of odds ratios and hazard ratios to risk ratios},
-  author={VanderWeele, Tyler J},
-  journal={Biometrics},
-  volume={76},
-  number={3},
-  pages={746--752},
-  year={2020},
-  publisher={Wiley Online Library}
-}
-@book{lash2009applying,
-  title={Applying quantitative bias analysis to epidemiologic data},
-  author={Lash, Timothy L and Fox, Matthew P and Fink, Aliza K and others},
-  volume={533},
-  year={2009},
-  publisher={Springer}
-}
-@article{o1992prenatal,
-  title={Prenatal events and the risk of subependymal/intraventricular haemorrhage in very low birthweight neonates},
-  author={O'Shea, Michael and Savitz, David A and Hage, Marvin L and Feinstein, KA},
[…]
