### 2022-06-08 a6d2ceb
remove extra examples to shorten the paper
