+++ paper.md
@@ -21,15 +21,7 @@ bibliography: paper.bib
 
 # Summary
 
-The strength of evidence provided by epidemiological and observational
-studies is inherently limited by the potential for unmeasured confounding.
-We focus on three key quantities: the observed bound of the confidence interval
-closest to the null, a plausible residual effect size for an unmeasured continuous
-or binary confounder, and a realistic mean difference or prevalence difference for
-this hypothetical confounder. Building on the methods put forth by
-@Cornfield, @Bross, @Schlesselman, @Rosenbaum:1983, @Lin, @lash2009applying, @rosenbaum1986dropping, @cinelli2020making, @VanderWeele:2017ki, and @Ding, we can use these quantities to
-assess how an unmeasured confounder may tip our result to insignificance, rendering the
-study inconclusive.
+The strength of evidence provided by epidemiological and observational studies is inherently limited by the potential for unmeasured confounding. We focus on three key quantities: the observed bound of the confidence interval closest to the null, the relationship between an unmeasured confounder and the outcome, for example a plausible residual effect size for an unmeasured continuous or binary confounder, and the relationship between an unmeasured confounder and the exposure, for example a realistic mean difference or prevalence difference for this hypothetical confounder between exposure groups. Building on the methods put forth by @Cornfield, @Bross, @Schlesselman, @Rosenbaum:1983, @Lin, @lash2009applying, @rosenbaum1986dropping, @cinelli2020making, @VanderWeele:2017ki, and @Ding, we can use these quantities to assess how an unmeasured confounder may tip our result to insignificance, rendering the study inconclusive.
 
 # Statement of need
 
@@ -178,31 +170,11 @@ An overview of the main functions in `tipr` is summarized in **Table 1**:
 # Tipping Point Example
 
 After fitting your model, you can determine the unmeasured confounder
-needed to tip your analysis. This unmeasured confounder is determined by
-two quantities, the association between the exposure and the unmeasured
-confounder (if the unmeasured confounder is continuous, this is
-indicated with `smd`, if binary, with `exposed_p` and `unexposed_p`),
-and the association between the unmeasured confounder and outcome
-`outcome_association`. Using `tipr`, we can fix one of these and solve
-for the other. Alternatively, we can fix both and solve for `n`, that
-is, how many unmeasured confounders of this magnitude would tip the
-analysis.
+needed to tip your analysis. In this example, a model was fit and the exposure-outcome relationship was a relative risk of 1.5 (95% CI: 1.2, 1.8).
 
-In this example, a model was fit and the exposure-outcome relationship
-was a relative risk of 1.5 (95% CI: 1.2, 1.8).
+We are interested in a Normally distributed unmeasured confounder, so we can use the `tip_rr_with_continuous()` function. The function `tip()` is an alias for this function.
 
-## Continuous unmeasured confounder example
-
-We are interested in a Normally distributed unmeasured confounder, so we can use
-the `tip_rr_with_continuous()` function. The function `tip()` is an alias for 
-for this function (assuming the relationship you want to examine is the relative
-risk and the unmeasured confounder is Normally distributed.
-
-Let’s assume the relationship between the unmeasured confounder and
-outcome is 1.5 (`outcome_association = 1.5`), let's solve for the
-association between the exposure and unmeasured confounder needed to tip
-the analysis (in this case, we are solving for `smd`, the mean
-difference needed between the exposed and unexposed).
[…]
