### 2022-06-08 #4384 @fabian-s
.. and this checks the DOIs
