+++ R/tip_coef.R
@@ -94,7 +94,7 @@ tip_coef_one <- function(b, smd, outcome_association, verbose) {
     }
   }
   o <- tibble::tibble(
-    observed_effect = b,
+    effect_observed = b,
     smd = smd,
     outcome_association = outcome_association,
     n_unmeasured_confounders = n_unmeasured_confounders
@@ -103,7 +103,7 @@ tip_coef_one <- function(b, smd, outcome_association, verbose) {
     if (all(o$n_unmeasured_confounders == 0)) {
       o_notip <- o[o$n_unmeasured_confounders == 0,]
       message_glue(
-        "The observed effect ({round(o_notip$observed_effect, 2)}) ",
+        "The observed effect ({round(o_notip$effect_observed, 2)}) ",
         "cannot be tipped by an unmeasured confounder\nwith the ",
         "following specifications:",
         "\n  * estimated difference in scaled means between the ",
@@ -115,7 +115,7 @@ tip_coef_one <- function(b, smd, outcome_association, verbose) {
     } else if (any(o$n_unmeasured_confounders == 0)) {
       o_notip <- o[o$n_unmeasured_confounders == 0,]
       message_glue(
-        "The observed effect ({round(o_notip$observed_effect, 2)}) ",
+        "The observed effect ({round(o_notip$effect_observed, 2)}) ",
         "cannot be tipped by an unmeasured confounder\nwith the ",
         "following specifications:",
         "\n  * estimated difference in scaled means between the ",
@@ -127,7 +127,7 @@ tip_coef_one <- function(b, smd, outcome_association, verbose) {
 
       o_tip <- o[o$n_unmeasured_confounders != 0,]
       message_glue(
-        "The observed effect ({round(o_tip$observed_effect, 2)}) WOULD ",
+        "The observed effect ({round(o_tip$effect_observed, 2)}) WOULD ",
         "be tipped by {round(o$n_unmeasured_confounders)} ",
         "unmeasured confounder{ifelse(o_tip$n_unmeasured_confounders > 1, 's', '')}\n",
         "with the following specifications:",
@@ -139,7 +139,7 @@ tip_coef_one <- function(b, smd, outcome_association, verbose) {
       )
     } else {
       message_glue(
-        "The observed effect ({round(o$observed_effect, 2)}) WOULD ",
+        "The observed effect ({round(o$effect_observed, 2)}) WOULD ",
         "be tipped by {round(o$n_unmeasured_confounders)} ",
         "unmeasured confounder{ifelse(o$n_unmeasured_confounders > 1, 's', '')}\n",
         "with the following specifications:",
