+++ DESCRIPTION
@@ -1,7 +1,7 @@
 Package: tipr
 Type: Package
 Title: Tipping Point Analyses
-Version: 0.4.1
+Version: 0.4.2
 Author: Lucy D'Agostino McGowan
 Maintainer: Lucy D'Agostino McGowan <lucydagostino@gmail.com>
 Description: The strength of evidence provided by epidemiological and observational 
