### 2022-06-09 e847ece
shorten the paper a bit
