+++ paper.md
@@ -80,6 +80,8 @@ library(tipr)
 adjust_coef(effect = 1.5,
             smd = 0.1,
             outcome_association = 1)
+```
+
     ## The observed effect (1.5) is updated to 1.4 by a confounder with the 
     ## following specifications:
     ##  * estimated difference in scaled means: 0.1
@@ -89,7 +91,7 @@ adjust_coef(effect = 1.5,
     ## effect_adjusted effect_observed   smd outcome_association
     ##             <dbl>           <dbl> <dbl>               <dbl>
     ## 1             1.4             1.5   0.1                   1
-```
+
 
 # Functions
 
