### 2022-06-22 #4495 @editorialbot
[REVIEW]: tipr: An R package for sensitivity analyses for unmeasured confounders
**Submitting author:** @LucyMcGowan (<a href="http://orcid.org/0000-0001-7297-9359">Lucy D&#39;Agostino McGowan</a>)
**Repository:** https://github.com/LucyMcGowan/tipr
**Branch with paper.md** (empty if default branch): joss
**Version:** v1.0.0
