### 2022-06-29 #4495 @MichaelSchomaker
- The dependencies are not explicitly stated in the paper.
- The License file needs more information.
- There are no details on how to report issues/problems with the software

Comments about content: while the space in JOSS is limited, I have a few smaller suggestions and questions that could be addressed in the paper in 1-2 sentences each:
- is the method also applicable for more complex back-door paths, e.g. A <- L_1 -> L_2 -> Y? Or, to frame it differently, do the calculations require direct effects from confounder on outcome and confounder on exposure, respectively, and how is the method applicable to confounder-exposure/outcome relationships that are mediated through other variables. If there needs to be a direct effect relationship, it should be mentioned.
- the package contains 3 types of functions, the second for "Functions that calculate the magnitude of an unmeasured confounder needed to tip an observed effect to cross the null, rendering it inconclusive". In the abstract, the author mentions "we can use these quantities to assess how an unmeasured confounder may tip our result to insignificance, rendering the study inconclusive." The wording and functions represent a dichotomous viewpoint that may be similarly dangerous as  binary p-value cutoffs or looking  at overlaps of lower/upper confidence limits with null values. A CI shows the values of the parameter which, when combined with the background assumptions, produce a test model which is highly compatible with the data in the sense of having less than a certain amount of bits of information against it, see for example Rafi and Greenland: Semantic and cognitive tools to aid statistical science: replace confidence and significance by compatibility and surprise. This issue is also discussed in the original van der Weele article which is a main reference in this JOSS article: "We do not propose any threshold cutoff for the E-value. Enough mischief has been done by the arbitrary 0.05 P-value cutoff. The E-value, like the P-value, is a continuous measure. " Long story short: I would in abstract and function description emphasize the continuous nature of the approach and the potential danger of applying the (second set of) functions in way that looks only at when "an observed effect to cross the null, rendering it inconclusive". If I had the choice, I would actually strongly  favor an E-value example over the tipping point example.
- I would clearly define standardized mean difference, to avoid misunderstandings and misinterpretations.
- Other than that, I think this is a really useful R package with a well-written summary.
