+++ pkgdown.yml
@@ -1,6 +1,6 @@
 pandoc: 2.7.3
-pkgdown: 2.0.3
+pkgdown: 2.0.6
 pkgdown_sha: ~
 articles: {}
-last_built: 2022-06-09T12:44Z
+last_built: 2022-08-02T13:51Z
 
