+++ README.Rmd
@@ -76,3 +76,6 @@ if (requireNamespace("broom", quietly = TRUE) &&  requireNamespace("dplyr", quie
 }
 ```
 
+## Code of Conduct
+  
+Please note that the tipr project is released with a [Contributor Code of Conduct](https://contributor-covenant.org/version/2/0/CODE_OF_CONDUCT.html). By contributing to this project, you agree to abide by its terms.
