### 2022-08-02 9689054
update paper based on JOSS review
