+++ paper.md
@@ -25,11 +25,11 @@ The strength of evidence provided by epidemiological and observational studies i
 
 # Statement of need
 
-When assessing the relationship between an exposure and an outcome, the "no unmeasured confounders" assumption is crucial.[@rubin1974estimating; @DAgostino:1998tu] This assumption, however, is untestable, rendering sensitivity analyses necessary to quantify the potential impact of an unmeasured confounder. There are several related methods for conducting sensitivity analyses for unmeasured confounders[@Cornfield; @Bross; @Schlesselman; @Rosenbaum:1983; @Lin; @lash2009applying; @rosenbaum1986dropping; @cinelli2020making; @VanderWeele:2017ki; @Ding], however there is not currently a single R package that has a unified grammar allowing the user to conduct appropriate sensitivity analysis for their study. A unified grammar allows the user to easily transition between methods depending on the scenario under which they are conducting the sensitivity analysis as well as compare the results between methods. `tipr` seeks to fill this need.
+When assessing the relationship between an exposure and an outcome, the "no unmeasured confounders" assumption is crucial.[@rubin1974estimating; @DAgostino:1998tu] This assumption, however, is untestable, rendering sensitivity analyses necessary to quantify the potential impact of unmeasured confounding. There are several related methods for conducting sensitivity analyses for unmeasured confounders[@Cornfield; @Bross; @Schlesselman; @Rosenbaum:1983; @Lin; @lash2009applying; @rosenbaum1986dropping; @cinelli2020making; @VanderWeele:2017ki; @Ding], however there is not currently a single R package that has a unified grammar allowing the user to conduct appropriate sensitivity analysis for their study. A unified grammar allows the user to easily transition between methods depending on the scenario under which they are conducting the sensitivity analysis as well as compare the results between methods. `tipr` seeks to fill this need.
 
 # Overview 
 
-The `tipr` R package [@rstats] allows the user to conduct sensitivity analyses for unmeasured confounders. The user provides an "observed effect" (that is, some effect between an exposure of interest and an outcome of interest after adjusting for any observed confounders), and the sensitivity analyses will quantify how sensitive the effect is to a potential unmeasured confounder.  The `tipr` functions fall into three categories:
+The `tipr` R package [@rstats] allows the user to conduct sensitivity analyses for unmeasured confounding. The user provides an "observed effect" (that is, some effect between an exposure of interest and an outcome of interest after adjusting for any observed confounders), and the sensitivity analyses will quantify how sensitive the effect is to a potential unmeasured confounder.  The `tipr` functions fall into three categories:
 
 (1) Functions that calculate how an observed effects would change with a specified unmeasured confounder (`adjust` functions)
 (2) Functions that calculate the magnitude of an unmeasured confounder needed to tip an observed effect to cross the null, rendering it inconclusive (`tip` functions)
@@ -61,11 +61,11 @@ Informally, a confounder is a pre-exposure variable that is associated with both
 
 ## Unmeasured confounder - exposure relationship
 
[…]
