+++ R/tip.R
@@ -2,7 +2,7 @@
 #'
 #' choose one of the following, and the other will be estimated:
 #' * `smd`
-#' * `outcome_association`
+#' * `outcome_effect`
 #'
 #' @param effect Numeric positive value. Observed exposure - outcome effect
 #'    (assumed to be the exponentiated coefficient, so a relative risk, odds
@@ -10,7 +10,7 @@
 #'    bound, or upper confidence bound.
 #' @param smd Numeric. Estimated difference in scaled means between the
 #'    unmeasured confounder in the exposed population and unexposed population
-#' @param outcome_association Numeric positive value. Estimated association
+#' @param outcome_effect Numeric positive value. Estimated relationship
 #'    between the unmeasured confounder and the outcome
 #' @param verbose Logical. Indicates whether to print informative message.
 #'    Default: `TRUE`
@@ -26,13 +26,13 @@
 #' @return Data frame.
 #'
 #' @examples
-#' ## to estimate the association between an unmeasured confounder and outcome
+#' ## to estimate the relationship between an unmeasured confounder and outcome
 #' ## needed to tip analysis
 #' tip(1.2, smd = -2)
 #'
 #' ## to estimate the number of unmeasured confounders specified needed to tip
 #' ## the analysis
-#' tip(1.2, smd = -2, outcome_association = .99)
+#' tip(1.2, smd = -2, outcome_effect = .99)
 #'
 #' ## Example with broom
 #' if (requireNamespace("broom", quietly = TRUE) &&
@@ -41,19 +41,19 @@
 #'    broom::tidy(conf.int = TRUE, exponentiate = TRUE) %>%
 #'    dplyr::filter(term == "mpg") %>%
 #'    dplyr::pull(conf.low) %>%
-#'    tip(outcome_association = 2.5)
+#'    tip(outcome_effect = 2.5)
 #'}
 #' @export
-tip <- function(effect, smd = NULL, outcome_association = NULL,
+tip <- function(effect, smd = NULL, outcome_effect = NULL,
                 verbose = TRUE, correction_factor = "none") {
 
   smd <- smd %||% list(NULL)
-  outcome_association <- outcome_association %||% list(NULL)
+  outcome_effect <- outcome_effect %||% list(NULL)
 
   o <- purrr::pmap(
     list(b = effect,
          smd = smd,
-         outcome_association = outcome_association,
+         outcome_effect = outcome_effect,
          verbose = verbose,
          correction_factor = correction_factor),
     tip_one
@@ -61,48 +61,48 @@ tip <- function(effect, smd = NULL, outcome_association = NULL,
   do.call(rbind, o)
 }
 
-tip_one <- function(b, smd, outcome_association, verbose, correction_factor) {
+tip_one <- function(b, smd, outcome_effect, verbose, correction_factor) {
   check_effect(b)
-  check_gamma(outcome_association)
+  check_gamma(outcome_effect)
 
   correction <- ""
   if (correction_factor == "hr") {
     b <- hr_transform(b)
-    outcome_association <- hr_transform(outcome_association)
+    outcome_effect <- hr_transform(outcome_effect)
     correction <- 'You opted to use the hazard ratio correction to convert your hazard ratios to approximate risk ratios.\nThis is a good idea if the outcome is common (>15%).'
   }
 
   if (correction_factor == "or") {
     b <- or_transform(b)
-    outcome_association <- or_transform(outcome_association)
+    outcome_effect <- or_transform(outcome_effect)
     correction <- 'You opted to use the odds ratio correction to convert your odds ratios to approximate risk ratios.\nThis is a good idea if the outcome is common (>15%).'
   }
 
   n_unmeasured_confounders <- 1
 
-  if (is.null(outcome_association)) {
[…]
