+++ README.md
@@ -25,13 +25,13 @@ library(tipr)
 
 After fitting your model, you can determine the unmeasured confounder
 needed to tip your analysis. This unmeasured confounder is determined by
-two quantities, the association between the exposure and the unmeasured
+two quantities, the relationship between the exposure and the unmeasured
 confounder (if the unmeasured confounder is continuous, this is
 indicated with `smd`, if binary, with `exposed_p` and `unexposed_p`),
-and the association between the unmeasured confounder and outcome
-`outcome_association`. Using this 📦, we can fix one of these and solve
-for the other. Alternatively, we can fix both and solve for `n`, that
-is, how many unmeasured confounders of this magnitude would tip the
+and the relationship between the unmeasured confounder and outcome
+`outcome_effect`. Using this 📦, we can fix one of these and solve for
+the other. Alternatively, we can fix both and solve for `n`, that is,
+how many unmeasured confounders of this magnitude would tip the
 analysis.
 
 In this example, a model was fit and the exposure-outcome relationship
@@ -43,27 +43,27 @@ We are interested in a continuous unmeasured confounder, so we will use
 the `tip_with_continuous()` function.
 
 Let’s assume the relationship between the unmeasured confounder and
-outcome is 1.5 (`outcome_association = 1.5`), let’s solve for the
-association between the exposure and unmeasured confounder needed to tip
-the analysis (in this case, we are solving for `smd`, the mean
+outcome is 1.5 (`outcome_effect = 1.5`), let’s solve for the
+relationship between the exposure and unmeasured confounder needed to
+tip the analysis (in this case, we are solving for `smd`, the mean
 difference needed between the exposed and unexposed).
 
 ``` r
-tip(1.2, outcome_association = 1.5)
+tip(1.2, outcome_effect = 1.5)
 ```
 
     ## The observed effect (1.2) WOULD be tipped by 1 unmeasured confounder
     ## with the following specifications:
     ##   * estimated difference in scaled means between the unmeasured confounder
     ##     in the exposed population and unexposed population: 0.45
-    ##   * estimated association between the unmeasured confounder and the outcome: 1.5
+    ##   * estimated relationship between the unmeasured confounder and the outcome: 1.5
 
     ## # A tibble: 1 × 5
-    ##   effect_adjusted effect_observed   smd outcome_association n_unmeasured_confou…
-    ##             <dbl>           <dbl> <dbl>               <dbl>                <dbl>
-    ## 1               1             1.2 0.450                 1.5                    1
+    ##   effect_adjusted effect_observed   smd outcome_effect n_unmeasured_confounders
+    ##             <dbl>           <dbl> <dbl>          <dbl>                    <dbl>
+    ## 1               1             1.2 0.450            1.5                        1
 
-A hypothetical unobserved continuous confounder that has an association
+A hypothetical unobserved continuous confounder that has a relationship
 of 1.5 with the outcome would need a scaled mean difference between
 exposure groups of `0.45` to tip this analysis at the 5% level,
 rendering it inconclusive.
@@ -75,9 +75,9 @@ use the `tip_with_binary()` function.
 
 Let’s assume the unmeasured confounder is prevalent in 25% of the
 exposed population (`exposed_p = 0.25`) and in 10% of the unexposed
-population (`unexposed_p = 0.10`) – let’s solve for the association
+population (`unexposed_p = 0.10`) – let’s solve for the relationship
 between the unmeasured confounder and the outcome needed to tip the
-analysis (`outcome_association`).
+analysis (`outcome_effect`).
 
 ``` r
[…]
