### 2022-08-02 00d319c
add CRAN instructions
