### 2022-08-02 e50840e
add CMD check via Github Actions
