+++ R/adjust_coefficient.R
@@ -1,13 +1,13 @@
 #' Adjust an observed regression coefficient for a normally distributed
 #' confounder
 #'
-#' @param effect Numeric. Observed exposure - outcome effect from a regression
+#' @param effect_observed Numeric. Observed exposure - outcome effect from a regression
 #'    model. This can be the beta coefficient, the lower confidence bound of
 #'    the beta coefficient, or the upper confidence bound of the beta
 #'    coefficient.
-#' @param smd Numeric. Estimated difference in scaled means between the
+#' @param exposure_confounder_effect Numeric. Estimated difference in scaled means between the
 #'    unmeasured confounder in the exposed population and unexposed population
-#' @param outcome_effect Numeric. Estimated relationship
+#' @param confounder_outcome_effect Numeric. Estimated relationship
 #'    between the unmeasured confounder and the outcome.
 #' @param verbose Logical. Indicates whether to print informative message.
 #'    Default: `TRUE`
@@ -20,30 +20,35 @@
 #' ## with a difference in scaled means between exposure groups of 0.2
 #' ## and coefficient of 0.3
 #' adjust_coef(0.5, 0.2, 0.3)
-adjust_coef <- function(effect, smd, outcome_effect, verbose = TRUE) {
-  effect_adj <- effect - outcome_effect * smd
-  o <- tibble::tibble(
-    effect_adjusted = effect_adj,
-    effect_observed = effect,
-    smd = smd,
-    outcome_effect = outcome_effect
-  )
-  if (verbose) {
-        message_glue(
-          "The observed effect ({round(effect, 2)}) ",
-          "is updated to {round(effect_adj, 2)} ",
-          "by a confounder with the following specifications:",
-          "\n  * estimated difference in scaled means: {smd}",
-          "\n  * estimated relationship between the unmeasured confounder and the ",
-          "outcome: {outcome_effect}\n"
-        )
+adjust_coef <-
+  function(effect_observed,
+           exposure_confounder_effect,
+           confounder_outcome_effect,
+           verbose = TRUE) {
+    effect_adj <-
+      effect_observed - confounder_outcome_effect * exposure_confounder_effect
+    o <- tibble::tibble(
+      effect_adjusted = effect_adj,
+      effect_observed = effect_observed,
+      exposure_confounder_effect = exposure_confounder_effect,
+      confounder_outcome_effect = confounder_outcome_effect
+    )
+    if (verbose) {
+      message_glue(
+        "The observed effect ({round(effect_observed, 2)}) ",
+        "is updated to {round(effect_adj, 2)} ",
+        "by a confounder with the following specifications:",
+        "\n  * estimated difference in scaled means: {exposure_confounder_effect}",
+        "\n  * estimated relationship between the unmeasured confounder and the ",
+        "outcome: {confounder_outcome_effect}\n"
+      )
+    }
+    return(o)
   }
-  return(o)
-}
 
 #' Adjust an observed coefficient from a loglinear model with a binary confounder
 #'
-#' @param effect Numeric. Observed exposure - outcome effect from a loglinear
+#' @param effect_observed Numeric. Observed exposure - outcome effect from a loglinear
 #'    model. This can be the beta coefficient, the lower confidence bound of
 #'    the beta coefficient, or the upper confidence bound of the beta
 #'    coefficient.
@@ -51,7 +56,7 @@ adjust_coef <- function(effect, smd, outcome_effect, verbose = TRUE) {
 #'    unmeasured confounder in the exposed population
[…]
