+++ R/tip_with_binary.R
@@ -2,23 +2,23 @@
 #'
 #' @description
 #' Choose two of the following three to specify, and the third will be estimated:
-#' * `exposed_p`
-#' * `unexposed_p`
-#' * `outcome_effect`
+#' * `exposed_confounder_prev`
+#' * `unexposed_confounder_prev`
+#' * `confounder_outcome_effect`
 #'
 #' Alternatively, specify all three and the function will return the number of unmeasured
 #' confounders specified needed to tip the analysis.
 #'
 #' @details [`tip_b()`] is an alias for [`tip_with_binary()`].
-#' @param effect Numeric positive value. Observed exposure - outcome effect
+#' @param effect_observed Numeric positive value. Observed exposure - outcome effect
 #'    (assumed to be the exponentiated coefficient, so a relative risk, odds
 #'    ratio, or hazard ratio). This can be the point estimate, lower confidence
 #'    bound, or upper confidence bound.
-#' @param exposed_p Numeric between 0 and 1. Estimated prevalence of the
+#' @param exposed_confounder_prev Numeric between 0 and 1. Estimated prevalence of the
 #'    unmeasured confounder in the exposed population
-#' @param unexposed_p Numeric between 0 and 1. Estimated prevalence of the
+#' @param unexposed_confounder_prev Numeric between 0 and 1. Estimated prevalence of the
 #'    unmeasured confounder in the unexposed population
-#' @param outcome_effect Numeric positive value. Estimated relationship
+#' @param confounder_outcome_effect Numeric positive value. Estimated relationship
 #'    between the unmeasured confounder and the outcome
 #' @param verbose Logical. Indicates whether to print informative message.
 #'    Default: `TRUE`
@@ -33,14 +33,14 @@
 #' @examples
 #' ## to estimate the relationship between an unmeasured confounder and outcome
 #' ## needed to tip analysis
-#' tip_with_binary(1.2, exposed_p = 0.5, unexposed_p = 0)
+#' tip_with_binary(1.2, exposed_confounder_prev = 0.5, unexposed_confounder_prev = 0)
 #'
 #' ## to estimate the number of unmeasured confounders specified needed to tip
 #' ## the analysis
 #' tip_with_binary(1.2,
-#'   exposed_p = 0.5,
-#'   unexposed_p = 0,
-#'   outcome_effect = 1.1)
+#'   exposed_confounder_prev = 0.5,
+#'   unexposed_confounder_prev = 0,
+#'   confounder_outcome_effect = 1.1)
 #'
 #' ## Example with broom
 #' if (requireNamespace("broom", quietly = TRUE) &&
@@ -49,24 +49,24 @@
 #'    broom::tidy(conf.int = TRUE, exponentiate = TRUE) %>%
 #'    dplyr::filter(term == "mpg") %>%
 #'    dplyr::pull(conf.low) %>%
-#'    tip_with_binary(exposed_p = 1, outcome_effect = 1.15)
+#'    tip_with_binary(exposed_confounder_prev = 1, confounder_outcome_effect = 1.15)
 #'}
 #' @export
-tip_with_binary <- function(effect,
-                            exposed_p = NULL,
-                            unexposed_p = NULL,
-                            outcome_effect = NULL,
+tip_with_binary <- function(effect_observed,
+                            exposed_confounder_prev = NULL,
+                            unexposed_confounder_prev = NULL,
+                            confounder_outcome_effect = NULL,
                             verbose = TRUE,
                             correction_factor = "none") {
-  exposed_p <- exposed_p %||% list(NULL)
-  unexposed_p <- unexposed_p %||% list(NULL)
-  outcome_effect <- outcome_effect %||% list(NULL)
[…]
