+++ R/tip-helpers.R
@@ -141,26 +141,50 @@ check_r2 <- function(r2, exposure = FALSE, effect, se, df) {
   }
   if (exposure) {
     if (any(r2 == 1)) {
-      stop_glue("You input:\n * `exposure_r2`: {r2}\n",
-                "This means 100% of the residual variation in the exposure ",
-                "is explained by the unmeasured confounder, meaning regardless ",
-                "of the unmeasured confounder - outcome relationship, this ",
-                "will be \"tipped\".")
+      stop_glue(
+        "You input:\n * `exposure_r2`: {r2}\n",
+        "This means 100% of the residual variation in the exposure ",
+        "is explained by the unmeasured confounder, meaning regardless ",
+        "of the unmeasured confounder - outcome relationship, this ",
+        "will be \"tipped\"."
+      )
     }
     limit <- sensemakr::partial_r2(effect / se, df)
     if (any(r2 < limit)) {
-      stop_glue("You input:\n * `exposure_r2`: {r2[r2 < limit]}\n",
-                "It is not possible to tip this result with any unmeasured ",
-                "confounder - outcome relationship. In fact, if your ",
-                "unmeasured confounder explained 100% of the residual ",
-                "variation in your outcome, the partial R2 for the unmeasured ",
-                "confounder - exposure relationship would have to be ",
-                "{round(limit, 3)} for the exposure - outcome relationship ",
-                "to be explained away.")
+      stop_glue(
+        "You input:\n * `exposure_r2`: {r2[r2 < limit]}\n",
+        "It is not possible to tip this result with any unmeasured ",
+        "confounder - outcome relationship. In fact, if your ",
+        "unmeasured confounder explained 100% of the residual ",
+        "variation in your outcome, the partial R2 for the unmeasured ",
+        "confounder - exposure relationship would have to be ",
+        "{round(limit, 3)} for the exposure - outcome relationship ",
+        "to be explained away."
+      )
     }
   }
+}
+tip_exposure_r2 <- function(effect, se, df, outcome_r2) {
+  if (is.null(outcome_r2)) {
+    stop(
+      "Please input at least one of the following:\n * `exposure_r2`\n * `outcome_r2`",
+      call. = FALSE
+    )
   }
-  tip_exposure_r2 <- function(effect, se, df, outcome_r2) {
+  check_r2(outcome_r2)
+
+  exposure_r2 <-
+    effect ^ 2 / (effect ^ 2 + se ^ 2 * df * outcome_r2)
+  if (any(exposure_r2 > 1)) {
+    stop_glue(
+      "Given the input:\n * `effect`: {effect}\n * `outcome_r2`: {outcome_r2[exposure_r2 > 1]}\n",
+      "There does not exist an unmeasured confounder that could tip this.\n",
+    )
+  }
+  as.numeric(exposure_r2)
+}
+tip_exposure_r2_bound <-
+  function(effect, se, df, outcome_r2, alpha) {
     if (is.null(outcome_r2)) {
[…]

[…]
