### 2022-08-02 12ad5e6
update readme with new API
