+++ README.md
@@ -37,11 +37,12 @@ After fitting your model, you can determine the unmeasured confounder
 needed to tip your analysis. This unmeasured confounder is determined by
 two quantities, the relationship between the exposure and the unmeasured
 confounder (if the unmeasured confounder is continuous, this is
-indicated with `smd`, if binary, with `exposed_p` and `unexposed_p`),
-and the relationship between the unmeasured confounder and outcome
-`outcome_effect`. Using this 📦, we can fix one of these and solve for
-the other. Alternatively, we can fix both and solve for `n`, that is,
-how many unmeasured confounders of this magnitude would tip the
+indicated with `exposure_confounder_effect`, if binary, with
+`exposed_confounder_prev` and `unexposed_confounder_prev`), and the
+relationship between the unmeasured confounder and outcome
+`confounder_outcome_effect`. Using this 📦, we can fix one of these and
+solve for the other. Alternatively, we can fix both and solve for `n`,
+that is, how many unmeasured confounders of this magnitude would tip the
 analysis.
 
 In this example, a model was fit and the exposure-outcome relationship
@@ -50,16 +51,18 @@ was 1.5 (95% CI: 1.2, 1.8).
 ## Continuous unmeasured confounder example
 
 We are interested in a continuous unmeasured confounder, so we will use
-the `tip_with_continuous()` function.
+the `tip_with_continuous()` function. The package comes with an example
+data set simulated for this scenario, `exdata_continuous`.
 
 Let’s assume the relationship between the unmeasured confounder and
-outcome is 1.5 (`outcome_effect = 1.5`), let’s solve for the
+outcome is 1.5 (`confounder_outcome_effect = 1.5`), let’s solve for the
 relationship between the exposure and unmeasured confounder needed to
-tip the analysis (in this case, we are solving for `smd`, the mean
-difference needed between the exposed and unexposed).
+tip the analysis (in this case, we are solving for
+`exposure_confounder_effect`, the mean difference needed between the
+exposed and unexposed).
 
 ``` r
-tip(1.2, outcome_effect = 1.5)
+tip(effect_observed = 1.2, confounder_outcome_effect = 1.5)
 ```
 
     ## The observed effect (1.2) WOULD be tipped by 1 unmeasured confounder
@@ -69,9 +72,10 @@ tip(1.2, outcome_effect = 1.5)
     ##   * estimated relationship between the unmeasured confounder and the outcome: 1.5
 
     ## # A tibble: 1 × 5
-    ##   effect_adjusted effect_observed   smd outcome_effect n_unmeasured_confounders
-    ##             <dbl>           <dbl> <dbl>          <dbl>                    <dbl>
-    ## 1               1             1.2 0.450            1.5                        1
+    ##   effect_adjusted effect_observed exposure_confounder_effect confounder_outcome…
+    ##             <dbl>           <dbl>                      <dbl>               <dbl>
+    ## 1               1             1.2                      0.450                 1.5
+    ## # … with 1 more variable: n_unmeasured_confounders <dbl>
 
 A hypothetical unobserved continuous confounder that has a relationship
 of 1.5 with the outcome would need a scaled mean difference between
@@ -84,13 +88,15 @@ Now we are interested in the binary unmeasured confounder, so we will
 use the `tip_with_binary()` function.
 
 Let’s assume the unmeasured confounder is prevalent in 25% of the
-exposed population (`exposed_p = 0.25`) and in 10% of the unexposed
-population (`unexposed_p = 0.10`) – let’s solve for the relationship
[…]
