+++ NEWS.md
@@ -10,6 +10,8 @@
 * `exposure_r2` -> `confounder_exposure_r2`
 * `outcome_r2` -> `confounder_outcome_r2`
 
+* Added two new example datasets: `exdata_continuous` and `exdata_rr`
+
 # tipr 0.4.2
 
 * Make the output tibble names consistent (`adjusted_effect` -> `effect_adjusted`)
