+++ README.Rmd
@@ -36,19 +36,63 @@ library(tipr)
 
 After fitting your model, you can determine the unmeasured confounder needed to tip your analysis. This unmeasured confounder is determined by two quantities, the relationship between the exposure and the unmeasured confounder (if the unmeasured confounder is continuous, this is indicated with `exposure_confounder_effect`, if binary, with `exposed_confounder_prev` and `unexposed_confounder_prev`), and the relationship between the unmeasured confounder and outcome `confounder_outcome_effect`. Using this `r emo::ji("package")`, we can fix one of these and solve for the other. Alternatively, we can fix both and solve for `n`, that is, how many unmeasured confounders of this magnitude would tip the analysis. 
 
-In this example, a model was fit and the exposure-outcome relationship was 1.5 (95% CI: 1.2, 1.8). 
+This package comes with a few example data sets. For this example, we will use `exdata_rr`. This data set was simulated such that there are two confounders, one that was "measured" (and thus useable in the main analysis, this is called `measured_confounder`) and one that is "unmeasured" (we have access to it because this is simulated data, but ordinarily we would not, this variable is called `.unmeasured_confounder`). 
+
+Using the example data `exdata_rr`, we can estimate the exposure-outcome relationship using the measured confounder as follows:
+
+```{r}
+mod <- glm(outcome ~ exposure + measured_confounder, data = exdata_rr, 
+           family = poisson)
+
+mod %>%
+  broom::tidy(exponentiate = TRUE, conf.int = TRUE)
+```
+
+We see the above example, the exposure-outcome relationship is 1.5 (95% CI: 1.09, 2.08). 
 
 ## Continuous unmeasured confounder example
 
-We are interested in a continuous unmeasured confounder, so we will use the `tip_with_continuous()` function. The package comes with an example data set simulated for this scenario, `exdata_continuous`. 
+We are interested in a continuous unmeasured confounder, so we will use the `tip_with_continuous()` function. 
 
-Let's assume the relationship between the unmeasured confounder and outcome is 1.5 (`confounder_outcome_effect = 1.5`),  let's solve for the relationship between the exposure and unmeasured confounder needed to tip the analysis (in this case, we are solving for `exposure_confounder_effect`, the mean difference needed between the exposed and unexposed).
+Let's assume the unmeaured confounder is normally distributed with a mean of 0.5 in the exposed group and 0 in the unexposed (and unit variance in both), resulting in a mean difference of 0.5 (`exposure_confounder_effect = 0.5`),  let's solve for the relationship between the unmeasured confounder and outcome needed to tip the analysis (in this case, we are solving for `confounder_outcome_effect`).
 
 ```{r}
-tip(effect_observed = 1.2, confounder_outcome_effect = 1.5)
+tip(effect_observed = 1.5, exposure_confounder_effect = 0.5)
 ```
 
-A hypothetical unobserved continuous confounder that has a relationship of 1.5 with the outcome would need a scaled mean difference between exposure groups of `0.45` to tip this analysis at the 5% level, rendering it inconclusive.
+A hypothetical unobserved continuous confounder a scaled mean difference between exposure groups of `0.5` would need a relationship of at least 2.25 with the outcome to tip this analysis at the point estimate.  
+
+```{r}
+tip(effect_observed = 1.09, exposure_confounder_effect = 0.5)
+```
+
+A hypothetical unobserved continuous confounder a scaled mean difference between exposure groups of `0.5` would need a relationship of at least 1.19 with the outcome to tip this analysis at the 5% level, rendering it inconclusive.
[…]
