+++ man/exdata_rr.Rd
@@ -1,9 +1,9 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/exdata_continuous.R
+% Please edit documentation in R/exdata_rr.R
 \docType{data}
-\name{exdata_continuous}
-\alias{exdata_continuous}
-\title{Example Data (Continuous Outcome)}
+\name{exdata_rr}
+\alias{exdata_rr}
+\title{Example Data (Relative Risk)}
 \format{
 A data frame with 2,000 rows and 4 columns:
 \itemize{
@@ -14,7 +14,7 @@ A data frame with 2,000 rows and 4 columns:
 }
 }
 \usage{
-exdata_continuous
+exdata_rr
 }
 \description{
 A data set simulated with two Normally distributed confounders, one
