### 2022-08-03 2ed429f
update paper based on review
Merge remote-tracking branch 'origin/master' into joss
