### 2022-08-03 #4495 @LucyMcGowan
Thank you so much for the reviews, they certainly helped make the paper better! Please find my point-by-point responses below.

# Reviewer 1

* In the repository readme it says that the license is MIT, but the license fail only contains copyright information (https://github.com/LucyMcGowan/tipr/blob/master/LICENSE).

*I have added a license.md file to the repository.*

* In the documentation on your website the CRAN installation could be mentioned. Did you actively decide not to do so?

*I have updated the README to include this.*

* It would be nice to run the tests with Travis of Github actions. There is a testsfolder and a testthat script using which the tests can be run. All tests passed.

*I have added the tests via Github actions*

* The paper includes an overview of all implemented functions and their respective functionality (Section Functions). According to the review criteria, the API should not be detailed in the paper, but in the software documentation. I would encourage to shorten or remove this section.

*I have removed this section and amended the previous to list the various methods (rather than all function names)*

* If you want to details the API in the paper (which is not encouraged), it would be helpful to have a table accompoagnying the text that explains each component in the grammar, i.e. what is smd, exposed_p, exposure_r2, ....

*I have removed this section*

* Other methods are mentioned, but the paper does not provide an overview over existing R packages and what they do.

*I have added the following text to describe the current (single package) landscape rather than an overview package*

> There are several related methods for conducting sensitivity analyses for unmeasured confounders [@Cornfield; @Bross; @Schlesselman; @Rosenbaum:1983; @Lin; @lash2009applying; @rosenbaum1986dropping; @cinelli2020making; @VanderWeele:2017ki; @Ding], some of which have their own R packages, for example methods in @cinelli2020making can be implemented using the `sensemakr` R package, `obsSens` implements methods from @Lin, and `EValue` implements methods in @VanderWeele:2017ki;. However, there is not currently a single R package that has a unified grammar allowing the user to conduct appropriate sensitivity analysis for their study.

* Spaces are missing before the citations (e.g. line 85), which does not seem to be the JOSS convention
Line 86, should be "for"

*I have corrected this spacing*

* In the paper, the functionality may be visualized with a figure, e.g. a simple causal graph that illustrates the aspects of the confounder that can be modified and an overview of the methods which can be used.

*I have added a DAG as well as an additional figure to the final example*

# Reviewer 2

* The dependencies are not explicitly stated in the paper.

*The dependencies are in the DESCRIPTION file of the package itself (and handled by CRAN on installation) I would be happy to include this information in the paper, however I couldn't find any other R papers that do -- is there a standard way to include this information?*

* The License file needs more information.

*I have added a license.md file to the repository*

* There are no details on how to report issues/problems with the software

*I have added a bugreports link to the DESCRIPTION file as well as a code of conduct to the repository (and linked in the README)*

Comments about content: while the space in JOSS is limited, I have a few smaller suggestions and questions that could be addressed in the paper in 1-2 sentences each:

* is the method also applicable for more complex back-door paths, e.g. A <- L_1 -> L_2 -> Y? Or, to frame it differently, do the calculations require direct effects from confounder on outcome and confounder on exposure, respectively, and how is the method applicable to confounder-exposure/outcome relationships that are mediated through other variables. If there needs to be a direct effect relationship, it should be mentioned.

*Mathematically, this is a bias adjustment, so would account for any direct effects (if there are several, correlated, confounders that are missing, you can think of this as measuring the summary score of the independent effects). I have changed the word "confounder" to "confounding" in several places to indicate that this is accounting for all "confounding". For example, the Normally distributed U could be the linear combination of several confounders.*

* the package contains 3 types of functions, the second for "Functions that calculate the magnitude of an unmeasured confounder needed to tip an observed effect to cross the null, rendering it inconclusive". In the abstract, the author mentions "we can use these quantities to assess how an unmeasured confounder may tip our result to insignificance, rendering the study inconclusive." The wording and functions represent a dichotomous viewpoint that may be similarly dangerous as binary p-value cutoffs or looking at overlaps of lower/upper confidence limits with null values. A CI shows the values of the parameter which, when combined with the background assumptions, produce a test model which is highly compatible with the data in the sense of having less than a certain amount of bits of information against it, see for example Rafi and Greenland: Semantic and cognitive tools to aid statistical science: replace confidence and significance by compatibility and surprise. This issue is also discussed in the original van der Weele article which is a main reference in this JOSS article: "We do not propose any threshold cutoff for the E-value. Enough mischief has been done by the arbitrary 0.05 P-value cutoff. The E-value, like the P-value, is a continuous measure. " Long story short: I would in abstract and function description emphasize the continuous nature of the approach and the potential danger of applying the (second set of) functions in way that looks only at when "an observed effect to cross the null, rendering it inconclusive". If I had the choice, I would actually strongly favor an E-value example over the tipping point example.

*Thank you for this important comment. I completely agree with the concern about dichotomizing. A main motivation for performing these tipping point analyses is that they suggest not only a different effect, but potentially a completely different **direction** of the effect itself, which could completely alter conclusions. I have added the following line to the summary:*

>The `adjust` functions allow an investigator to examine how a specific (or set of specific) confounders would change a result while the `tip` functions provide sensitivity analyses that allow an investigator to examine how extreme an unmeasured confounder would need to be in order to change the direction of the effect, and thus often the conclusions of the study.

* I would clearly define standardized mean difference, to avoid misunderstandings and misinterpretations.

*I have added the following text to specify what I mean by "standardized" as well as explained the difference in means*:

> If quantifying the impact of a standardized Normally distributed confounder, the impact of the unmeasured confounder on the exposure is parameterized as a difference in means between the unmeasured confounder in the exposed population and the unexposed population. By "standardized Normally distributed" we mean that the unmeasured confounder is Normally distributed with mean $\mu_1$ for the exposed and $\mu_0$ for the unexposed and unit variance. (Note a standardized Normally distributed confounder can be created from a Normally distributed confounder by dividing by the variance).

*I have also changed the name of the parameter since `smd` can have different meanings*

* Other than that, I think this is a really useful R package with a well-written summary.

*Thank you! Thank you again for reading through this paper, I really appreciate the comments.*
