### 2022-08-03 edceca4
make figure smaller, add table label
