### 2022-08-03 252a14d
fix code output to match analysis
