### 2022-08-03 bd2e67f
fix figure caption
