### 2022-08-03 #4495 @fabian-s
thank you @LucyMcGowan, excellent. I added some very minor comments on formatting / language directly in the diffs of your commits.

@MichaelSchomaker @gcskoenig are you satisfied with these changes as well?

re:
>
    The dependencies are not explicitly stated in the paper.

The dependencies are in the DESCRIPTION file of the package itself (and handled by CRAN on installation) I would be happy to include this information in the paper, however I couldn't find any other R papers that do -- is there a standard way to include this information?
