### 2022-08-03 bf33d6d
fix figure caption
