### 2022-08-03 ab59eeb
update readme, fix typo (thank you @fabian-s!)
