### 2022-08-03 #4495 @LucyMcGowan
Thank you so much! I tried to add a line space again and I think it fixed the caption?
