### 2022-08-03 8b64477
variance -> standard deviation
