+++ paper.md
@@ -61,7 +61,7 @@ Informally, a confounder is a pre-exposure variable that is associated with both
 
 ## Unmeasured confounder - exposure relationship
 
-If quantifying the impact of a standardized Normally distributed confounder, the impact of the unmeasured confounder on the exposure is parameterized as a difference in means between the unmeasured confounder in the exposed population and the unexposed population. By "standardized Normally distributed" we mean that the unmeasured confounder is Normally distributed with mean $\mu_1$ for the exposed and $\mu_0$ for the unexposed and unit variance. (Note a standardized Normally distributed confounder can be created from a Normally distributed confounder by dividing by the variance). This is specified using the `exposure_confounder_effect` parameter. If quantifying the impact of a binary confounder, the impact is quantified using two parameters, `exposed_confounder_prev`: The estimated prevalence of the unmeasured confounder in the exposed population and `unexposed_confounder_prev`: The estimated prevalence of the unmeasured confounder in the unexposed population. Alternatively, the user can choose to quantify the magnitude of the unmeasured confounder-exposure relationship in terms of the percent of variation in the exposure explained by the unmeasured confounder. This is specified using the `confounder_exposure_r2` parameter. 
+If quantifying the impact of a standardized Normally distributed confounder, the impact of the unmeasured confounder on the exposure is parameterized as a difference in means between the unmeasured confounder in the exposed population and the unexposed population. By "standardized Normally distributed" we mean that the unmeasured confounder is Normally distributed with mean $\mu_1$ for the exposed and $\mu_0$ for the unexposed and unit variance. (Note a standardized Normally distributed confounder can be created from a Normally distributed confounder by dividing by the standard deviation). This is specified using the `exposure_confounder_effect` parameter. If quantifying the impact of a binary confounder, the impact is quantified using two parameters, `exposed_confounder_prev`: The estimated prevalence of the unmeasured confounder in the exposed population and `unexposed_confounder_prev`: The estimated prevalence of the unmeasured confounder in the unexposed population. Alternatively, the user can choose to quantify the magnitude of the unmeasured confounder-exposure relationship in terms of the percent of variation in the exposure explained by the unmeasured confounder. This is specified using the `confounder_exposure_r2` parameter. 
 
 ## Unmeasured confounder - outcome relationship
 
