### 2022-08-03 #4495 @fabian-s
> Thank you so much! I tried to add a line space again and I think it fixed the caption?

looks like it :)

I think I'll just go ahead and make and editorial decision to accept this now , seems fine to me. (Sorry Gunnar & Michael, I want this done before my vacation)

@LucyMcGowan At this point could you please:
