### 2022-08-03 #4495 @LucyMcGowan
-[x]   Make a tagged release of your software, and list the version tag of the archived version here: https://github.com/LucyMcGowan/tipr/releases/tag/v1.0.0
-[x]   Archive the reviewed software in Zenodo or a similar service (e.g., figshare, an institutional repository): [10.5281/zenodo.6958926](https://zenodo.org/record/6958926#.YuqnZi-B0xU)
-[x]   Please list the DOI of the archived version here. 10.5281/zenodo.6958926
