### 2022-08-14 #4495 @MichaelSchomaker
Hi everyone, I have also been on vacation.

The paper is nice and the revisions good. Some last very minor points:

@LucyMcGowan: 1) you write: "if there are several, correlated, confounders that are missing, you can think of this as measuring the summary score of the independent effects" -> why not simply spelling this out in the paper?
2) why keeping the phrase "rendering it inconclusive" throughout the paper? The point estimate could cross the null, but the point estimates, together with the associated CI, may be similar suggestive of the direction of effect for different estimates/levels of unmeasured confounding. That is, the CI contains the set of parameter values that are more compatible with the data than those values lying outside the interval (given the background assumptions); a point estimate of -0.1 with a 95% CI of -0.3;0.1 may be similar suggestive of a positive effect as an estimate of 0.01 and CI of -0.05;0.07. I think it is an optional point, but you may consider either omitting the phrase or emphasizing more the descriptive an exploratory nature of the tool, rather than emphasizing the cutoff point.
