### 2022-08-20 #4495 @LucyMcGowan
@MichaelSchomaker I definitely hear what you’re saying re: “rendering it inconclusive” that language came from a review of the methods several years ago when we used to say “rendering it null” (since the interval included the null) and it was pointed out that just barely including the null was more likely inconclusive — but maybe I should just remove the “rendering…” text all together.

Technical question, @fabian-s I archived the previous version, if I made this change would I just increment that?
