+++ paper.bib
@@ -51,7 +51,7 @@ doi = {10.1111/j.2517-6161.1983.tb01242.x}
 }
 
 @article{Lin,
-author = {Lin, D Y and Psaty, B M and Kronmal, R A},
+author = {Lin, DY and Psaty, BM and Kronmal, RA},
 title = {{Assessing the sensitivity of regression results to unmeasured confounders in observational studies}},
 journal = {Biometrics},
 year = {1998},
@@ -93,7 +93,7 @@ doi = {10.7326/M16-2607}
 }
 
 @article{DAgostino:1998tu,
-author = {D'Agostino, R B},
+author = {D'Agostino, RB},
 title = {{Propensity score methods for bias reduction in the comparison of a treatment to a non-randomized control group}},
 journal = {Statistics in medicine},
 year = {1998},
@@ -114,7 +114,7 @@ doi = {10.1002/(SICI)1097-0258(19981015)17:19<2265::AID-SIM918>3.0.CO;2-B}
 }
 
 @article{rubin1974estimating,
-author = {Rubin, Donald B},
+author = {Rubin, DB},
 title = {{Estimating causal effects of treatments in randomized and nonrandomized studies}},
 journal = {Journal of educational Psychology},
 year = {1974},
@@ -149,7 +149,7 @@ doi = {10.1037/h0037350}
 
 @article{rosenbaum1986dropping,
   title={Dropping out of high school in the United States: An observational study},
-  author={Rosenbaum, Paul R},
+  author={Rosenbaum, PR},
   journal={Journal of Educational Statistics},
   volume={11},
   number={3},
