### 2022-08-30 #4495 @LucyMcGowan
Thank you! I have dropped the phrase from the current draft -- please let me know if you need any other updates from me. Thank you all for making this paper better!
