### 2022-08-31 1b4bc47
fix :bug: in adjust_*_with_binary functions
