+++ R/adjust_coefficient.R
@@ -52,9 +52,9 @@ adjust_coef <-
 #'    model. This can be the beta coefficient, the lower confidence bound of
 #'    the beta coefficient, or the upper confidence bound of the beta
 #'    coefficient.
-#' @param exposed_p Numeric between 0 and 1. Estimated prevalence of the
+#' @param exposed_confounder_prev Numeric between 0 and 1. Estimated prevalence of the
 #'    unmeasured confounder in the exposed population
-#' @param unexposed_p Numeric between 0 and 1. Estimated prevalence of the
+#' @param unexposed_confounder_prev Numeric between 0 and 1. Estimated prevalence of the
 #'    unmeasured confounder in the unexposed population
 #' @param confounder_outcome_effect Numeric. Estimated relationship
 #'    between the unmeasured confounder and the outcome.
@@ -68,23 +68,23 @@ adjust_coef <-
 #' adjust_coef_with_binary(1.1, 0.5, 0.3, 1.3)
 adjust_coef_with_binary <-
   function(effect_observed,
-           exposed_p,
-           unexposed_p,
+           exposed_confounder_prev,
+           unexposed_confounder_prev,
            confounder_outcome_effect,
            verbose = TRUE) {
-    check_prevalences(unexposed_p, exposed_p)
+    check_prevalences(unexposed_confounder_prev, exposed_confounder_prev)
 
-    confounding_factor <- log((exp(confounder_outcome_effect) * exposed_p + (1 - exposed_p)) /
+    confounding_factor <- log((exp(confounder_outcome_effect) * exposed_confounder_prev + (1 - exposed_confounder_prev)) /
                                 (
-                                  exp(confounder_outcome_effect) * unexposed_p + (1 - unexposed_p)
+                                  exp(confounder_outcome_effect) * unexposed_confounder_prev + (1 - unexposed_confounder_prev)
                                 ))
 
     effect_adj <- effect_observed - confounding_factor
     o <- tibble::tibble(
       effect_adjusted = effect_adj,
       effect_observed = effect_observed,
-      exposed_p = exposed_p,
-      unexposed_p = unexposed_p,
+      exposed_confounder_prev = exposed_confounder_prev,
+      unexposed_confounder_prev = unexposed_confounder_prev,
       confounder_outcome_effect = confounder_outcome_effect
     )
     if (verbose) {
@@ -93,8 +93,8 @@ adjust_coef_with_binary <-
         "is updated to {round(effect_adj, 2)} ",
         "by an confounder with the following specifications:",
         "\n  * estimated prevalence of the unmeasured confounder ",
-        "in the exposed population: {round(exposed_p, 2)}\n  * estimated prevalence of ",
-        "the unmeasured confounder in the unexposed population: {round(unexposed_p, 2)}",
+        "in the exposed population: {round(exposed_confounder_prev, 2)}\n  * estimated prevalence of ",
+        "the unmeasured confounder in the unexposed population: {round(unexposed_confounder_prev, 2)}",
         "\n  * estimated relationship between the unmeasured confounder and the ",
         "outcome: {round(confounder_outcome_effect, 2)}\n"
       )
@@ -283,9 +283,9 @@ adjust_or <-
 #' @param effect_observed Numeric positive value. Observed exposure - outcome relative risk.
 #'    This can be the point estimate, lower confidence bound, or upper
 #'    confidence bound.
-#' @param exposed_p Numeric between 0 and 1. Estimated prevalence of the
+#' @param exposed_confounder_prev Numeric between 0 and 1. Estimated prevalence of the
 #'    unmeasured confounder in the exposed population
-#' @param unexposed_p Numeric between 0 and 1. Estimated prevalence of the
[…]
