### 2023-10-06 #4495 @ajstewartlang
@LucyMcGowan Might you perhaps have the bandwidth to review the following submission please?

https://github.com/openjournals/joss-reviews/issues/5862
