### 2023-10-08 #4495 @arfon
@LucyMcGowan – you'll need to teach @editorialbot how to structure your name. Could you please update your name in your `paper.md` following this guide: https://joss.readthedocs.io/en/latest/submitting.html#article-metadata

@kthyng – then we'll need to do and `@editorialbot reaccept`
